factor out hits_from_hints to context/utils.py
