+++ tests/context/test_solr.py
@@ -179,12 +179,6 @@ def test_hints(ctx: SolrContext, index: str, query: Query):
     assert len(hints["variable_id"]) > 1
 
 
-def test_hits_from_hints(ctx: SolrContext):
-    hints = {"facet_name": {"value_a": 1, "value_b": 2, "value_c": 3}}
-    hits = ctx.hits_from_hints(hints)
-    assert hits == [6]
-
-
 @parametrized_index
 @pytest.mark.parametrize(
     "query_all",
